refactor: Use react-redux-firebase@next in example/material
