+++ examples/complete/material/package.json
@@ -7,6 +7,7 @@
     "clean": "rimraf dist",
     "start": "react-scripts start",
     "build": "react-scripts build",
+    "start:build": "npm run build && firebase serve --port=3000",
     "eject": "react-scripts eject",
     "build:config": "firebase-ci createConfig",
     "deploy": "firebase-ci deploy -s",
@@ -36,7 +37,7 @@
     "react-google-button": "^0.5.3",
     "react-loadable": "^5.5.0",
     "react-redux": "^5.0.7",
-    "react-redux-firebase": "^2.1.9",
+    "react-redux-firebase": "^3.0.0-alpha.11",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.1",
     "recompose": "^0.30.0",
