+++ examples/complete/material/src/containers/App/App.js
@@ -3,17 +3,50 @@ import PropTypes from 'prop-types'
 import { BrowserRouter as Router } from 'react-router-dom'
 import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles'
 import { Provider } from 'react-redux'
+import { ReactReduxFirebaseProvider } from 'react-redux-firebase'
 import ThemeSettings from 'theme'
-
+import firebase from 'firebase/app'
+import { firebase as fbConfig, reduxFirebase as rrfConfig } from '../../config'
 const theme = createMuiTheme(ThemeSettings)
 
-const App = ({ routes, store }) => (
-  <MuiThemeProvider theme={theme}>
-    <Provider store={store}>
-      <Router>{routes}</Router>
-    </Provider>
-  </MuiThemeProvider>
-)
+const App = ({ routes, store }) => {
+  // ======================================================
+  // Firebase Initialization
+  // ======================================================
+  firebase.initializeApp(fbConfig)
+
+  // ======================================================
+  // Redux + Firebase Config (react-redux-firebase & redux-firestore)
+  // ======================================================
+  const defaultRRFConfig = {
+    userProfile: 'users', // root that user profiles are written to
+    updateProfileOnLogin: false, // enable/disable updating of profile on login
+    presence: 'presence', // list currently online users under "presence" path in RTDB
+    sessions: null, // Skip storing of sessions
+    enableLogging: false // enable/disable Firebase Database Logging
+    // profileDecorator: (userData) => ({ email: userData.email }) // customize format of user profile
+  }
+
+  // Combine default config with overrides if they exist (set within .firebaserc)
+  const combinedConfig = rrfConfig
+    ? { ...defaultRRFConfig, ...rrfConfig }
+    : defaultRRFConfig
+
+  const rrfProps = {
+    firebase,
+    config: combinedConfig,
+    dispatch: store.dispatch
+  }
+  return (
+    <MuiThemeProvider theme={theme}>
+      <Provider store={store}>
+        <ReactReduxFirebaseProvider {...rrfProps}>
+          <Router>{routes}</Router>
+        </ReactReduxFirebaseProvider>
+      </Provider>
+    </MuiThemeProvider>
+  )
+}
 
 App.propTypes = {
   routes: PropTypes.object.isRequired,
