+++ examples/complete/material/src/store/createStore.js
@@ -1,35 +1,13 @@
 import { applyMiddleware, compose, createStore } from 'redux'
 import thunk from 'redux-thunk'
-import { reactReduxFirebase, getFirebase } from 'react-redux-firebase'
-import firebase from 'firebase/app'
+import { getFirebase } from 'react-redux-firebase'
 import 'firebase/database'
 import 'firebase/auth'
 import 'firebase/storage'
 import makeRootReducer from './reducers'
-import {
-  firebase as fbConfig,
-  reduxFirebase as rrfConfig,
-  env
-} from '../config'
+import { env } from '../config'
 
 export default (initialState = {}) => {
-  // ======================================================
-  // Redux + Firebase Config (react-redux-firebase & redux-firestore)
-  // ======================================================
-  const defaultRRFConfig = {
-    userProfile: 'users', // root that user profiles are written to
-    updateProfileOnLogin: false, // enable/disable updating of profile on login
-    presence: 'presence', // list currently online users under "presence" path in RTDB
-    sessions: null, // Skip storing of sessions
-    enableLogging: false // enable/disable Firebase Database Logging
-    // profileDecorator: (userData) => ({ email: userData.email }) // customize format of user profile
-  }
-
-  // Combine default config with overrides if they exist (set within .firebaserc)
-  const combinedConfig = rrfConfig
-    ? { ...defaultRRFConfig, ...rrfConfig }
-    : defaultRRFConfig
-
   // ======================================================
   // Store Enhancers
   // ======================================================
@@ -51,11 +29,6 @@ export default (initialState = {}) => {
   ]
 
   // ======================================================
-  // Firebase Initialization
-  // ======================================================
-  firebase.initializeApp(fbConfig)
-
-  // ======================================================
   // Store Instantiation and HMR Setup
   // ======================================================
   const store = createStore(
@@ -63,7 +36,6 @@ export default (initialState = {}) => {
     initialState,
     compose(
       applyMiddleware(...middleware),
-      reactReduxFirebase(firebase, combinedConfig),
       ...enhancers
     )
   )
